backend/requirements.txt

Django>=5.0,<6.0
djangorestframework>=3.14
django-environ>=0.11
django-cors-headers>=4.4
drf-spectacular>=0.27
pytest-django>=4.7
pytest>=7.4
ruff>=0.1
black>=23.12
isort>=5.13
